requirements.txt
arxiv
h5py
sentence-transformers
numpy
scikit-learn
rank-bm25
torch
python-dotenv
boto3
pymupdf
requests
tiktoken
nltk
rouge-score
sacrebleu
bert-score
openai
qdrant-client
fastapi
uvicorn
